KS-11-2 class User2, creating and invoking class methods and properties -- combine into one file
